Move all typescript stuff around so I can use correct filepaths when converting to ES6
