+++ async.js
@@ -2,7 +2,7 @@
  * Created by novacrazy on 1/17/2015.
  */
 
-module.exports = require( './build/async.js' );
+module.exports = require( './ts_build/async.js' );
 
 module.exports.Promise = require( 'bluebird' );
