+++ ts_build/common.js
