+++ ts_src/typings/bluebird/bluebird.d.ts
