+++ ts_src/typings/node/node.d.ts
